remove tests. all routes for google auth working
