+++ plugins/auth-backend/src/providers/google/index.ts
@@ -1,1 +1,1 @@
-// export { GoogleAuthProvider } from './provider';
+export { GoogleAuthProvider } from './provider';
